+++ sperl_hash_func.h
@@ -4,6 +4,6 @@
 #include "sperl_base.h"
 
 // Hash function
-int64_t SPerl_HASH_FUNC_calc_hash(SPerl* sperl, const char* str, int32_t len);
+int64_t SPerl_HASH_FUNC_calc_hash(SPerl* sperl, const char* str, int64_t len);
 
 #endif
